feat/2-navbar(rearrange pages)
