+++ pages/Portfolios.tsx
@@ -1,0 +1,13 @@
+import { NavBar } from "../components/NavBar";
+import { Title } from "../components/StyledComponents";
+
+const Portfolios = () => {
+  return (
+    <>
+      <NavBar></NavBar>
+      <Title>Your portfolios</Title>
+    </>
+  );
+};
+
+export default Portfolios;
